Структуры ДТО добавлены каждый в другой файл
