+++ Recepies/Modules/Recipes/Model/Category/CategoryType.swift
@@ -5,23 +5,23 @@ import Foundation
 
 /// Describe types of categories
 enum CategoryType: String, CaseIterable {
-    ///Sakad dishes
+    /// Sakad dishes
     case salad
-    ///Soup dishes
+    /// Soup dishes
     case soup
-    ///Chicken dishes
+    /// Chicken dishes
     case chicken
-    ///Meat dishes
+    /// Meat dishes
     case meat
-    ///Fish dishes
+    /// Fish dishes
     case fish
-    ///Side dishes
+    /// Side dishes
     case sideDish = "side Dish"
-    ///Panckake dishes
+    /// Panckake dishes
     case pancake
     /// Drinks
     case drinks
-    ///Dessert dishes
+    /// Dessert dishes
     case desserts
     /// Describe category name, used in Category model and search through RestAPI
     var description: String {
