+++ Recepies/Services/Network/DTO/CategoryDTO/RecipeDTO.swift
@@ -1,23 +1,6 @@
-// CategoryDTO.swift
+// RecipeDTO.swift
 // Copyright © RoadMap. All rights reserved.
 
-import Foundation
-
-// MARK: - CategoryDTO
-/// DTO for search by category
-struct CategoryDTO: Codable {
-    /// Array of recipesDTO containers
-    let hits: [Hit]
-}
-
-// MARK: - Hit
-/// RecipeDTO container
-struct Hit: Codable {
-    /// Recipe
-    let recipe: RecipeDTO
-}
-
-// MARK: - Recipe
 /// DTO for recipe
 struct RecipeDTO: Codable {
     /// Base link for detailed search
