+++ Recepies/Services/Network/DTO/DetailDTO/RecipeDTOContainer.swift
@@ -1,0 +1,8 @@
+// RecipeDTOContainer.swift
+// Copyright © RoadMap. All rights reserved.
+
+/// RecipeDTOContainer
+
+struct RecipeDTOContainer: Codable {
+    let recipe: RecipeDetailDTO
+}
